Seems I have been able to convert all functions to Swift 3.0 . Will see with the watch, I am not very sure and must study more.
has been easier than thought at the beginning and some updates to URL and check types with is instead ok kindOf are welcome. Less tests of optionals for the URL components.

Also very happy with the CoreDraw path and BeZier. Much cleaner interface.

Now the DEBUG works with everything. First connects to the wheel or service provider and analyzes all services/characteristics and records that in the log. Also presents all services and characteristics. probably will splinoff a new program for analyzing BLE connections but now I must integrate this knowledge in the normal connection to generate the correct adapter.