Changed the Track list organization. Instead of looking for files in the directory it maintains a file (tracks) and a dictionary. in memory with a little resume of the data. That makes it easier to present information and opens the way to present them in MAP form geographically and made searches.
So there are modifications in createPackage to add a new record,
and in ViewController to in sortFilesIntoSections to use the new database and also when commiting a delete.

Of course the big problem is when it desinchronizes. That is solved with a Rebuild Index option. It works in background and just updates the list when finished.
Also the new SimpleDatabadr and SimpleDatabaseObject are a simple database based in a Dictionary. Will change the WheelDatabase to that library in a future.

As the library is quite Object Agnostic and the interface is quite simple it allows the program to be moved to different databases.

Best point is that this system is the base for iCloud support in a future

Also modified the Calibration Recalculation system :

	- Added a reset button to set the values to 1.0 if nothing works :(
	- Moved the recalibration procedure to background
	- Added a progress indicator to show % or recalibration done. Now the wheel recalculation procedure accepts an optiona UIProgressView to show progress

Finally corrected bugs in differen modules, and in the translation files.